change max_tokens type to int

+++ litellm/types/completion.py
@@ -148,7 +148,7 @@ class CompletionRequest(BaseModel):
     n: Optional[int] = None
     stream: Optional[bool] = None
     stop: Optional[dict] = None
-    max_tokens: Optional[float] = None
+    max_tokens: Optional[int] = None
     presence_penalty: Optional[float] = None
     frequency_penalty: Optional[float] = None
     logit_bias: Optional[dict] = None
